# Mobx6 - Add action name in prod too