--- a/docs/best/pitfalls.md
+++ b/docs/best/pitfalls.md
@@ -209,7 +209,9 @@ export const MyComponent = observer(props => <div>hi</div>)
 myComponent.displayName = "MyComponent"
 
 // 2 (MobX infers component name from function name)
-export const MyComponent = observer(function MyComponent(props) { return <div>hi</div> })
+export const MyComponent = observer(function MyComponent(props) {
+    return <div>hi</div>
+})
 
 // 3 (transpiler will infer component name from variable name)
 const _MyComponent = props => <div>hi</div>