--- a/src/utils/eq.ts
+++ b/src/utils/eq.ts
@@ -83,7 +83,8 @@ function eq(a: any, b: any, depth: number, aStack?: any[], bStack?: any[]) {
                 typeof bCtor === "function" &&
                 bCtor instanceof bCtor
             ) &&
-            "constructor" in a && "constructor" in b
+            "constructor" in a &&
+            "constructor" in b
         ) {
             return false
         }