<!--
    If you are creating an extensive PR, you might want to open an issue with your idea first in case there is a chance for rejecting it.

    If you intend to work on PR over several days, please, create [draft pull requests](https://github.blog/2019-02-14-introducing-draft-pull-requests/).

    Keep in mind we are maintaining MobX 4 & 5 in the repo. If applicable, change both.

    Please look at the following checklist to ensure that your PR can be accepted quickly:
-->
## Changes:
- Define `action.name` unconditionally. Fixes #2263 
- Update readme to reflect correct scripts
- Fix `prettier` script to overwrite files
- Fix `.prettierignore` to ignore generated assets
- Run `prettier` on repo: 2 changed files

### Code change checklist

-   [x] Added/updated unit tests
-   [x] Updated changelog
-   [x] Updated `/docs`. For new functionality, at least `API.md` should be updated
-   [x] Verified that there is no significant performance drop (`npm run perf`)

<!--
    Feel free to ask help with any of these boxes!
-->

I'm not sure about how to test for perf regressions: I `yarn build && yarn perf` before and after with these results:

Before:
```shell
yarn test:performance proxy
Completed performance suite in 60.257 sec.
64.66user 1.62system 1:00.32elapsed 109%CPU (0avgtext+0avgdata 1092404maxresident)k
1136inputs+16outputs (1major+476988minor)pagefaults 0swaps

yarn test:performance legacy
Completed performance suite in 61.295 sec.
66.08user 1.66system 1:01.36elapsed 110%CPU (0avgtext+0avgdata 1033992maxresident)k
0inputs+16outputs (0major+477497minor)pagefaults 0swaps
```

After:
```shell
test/perf/index.js proxy
Completed performance suite in 61.343 sec.
64.80user 1.74system 1:01.41elapsed 108%CPU (0avgtext+0avgdata 1065372maxresident)k
0inputs+16outputs (0major+484730minor)pagefaults 0swaps

test/perf/index.js legacy
Completed performance suite in 60.269 sec.
65.21user 1.68system 1:00.33elapsed 110%CPU (0avgtext+0avgdata 1062128maxresident)k
0inputs+16outputs (0major+475698minor)pagefaults 0swaps
```
Looks pretty much the same.

<a href="https://gitpod.io/#https://github.com/mobxjs/mobx/pull/2343"><img src="https://gitpod.io/api/apps/github/pbs/github.com/elektronik2k5/mobx.git/9088bfd8a1990257b28512f7909f3eb222778ea3.svg" /></a>