@elektronik2k5 2020-04-25
Should I do a few more changes (is separate PRs)? Here's a list:
- Upgrade all deps
- Add spell checker
- Run tests against built package (not for perf)