@FredyC 2020-04-25
Thanks for doing it. I am not sure you should even do other unrelated changes you did here. @mweststrate is working on that branch and if he has some unpublished changes it could cause unnecessary conflicts.

I will leave merging up to him in case there are something he wants to do differently.