@mweststrate 2020-04-25
@elektronik2k5 

1. go ahead!
2. can you clarify? 
2. yes that would be awesome, not sure how to deal with obfuscated / not-in-prod errors. Maybe those tests need to be in an if block?