@mweststrate 2020-04-25
Unrelated, I was thinking making `observableRequiresReaction` the default, how'd you feel about that?