@elektronik2k5 2020-04-25
> 2. can you clarify?

Yes, here's a story: as a renown quality and grammar Nazi :smirk:, I review all of my client's code and catch a lot of typos. However, I once had to work with a team member who was particularly bad at spelling and got sick of pointing out obvious typos. So I decided to add a spell checker as a git hook and it worked well :). However, I also discovered that the codebase still has loads of typos which I missed.

The point is: a computer can spell check much better than any person. I've been using [cspell](https://github.com/streetsidesoftware/cspell) to catch spelling errors in codebases for a few years now.
It works exactly like any other linter for files staged for a commit: fail if non English words are found. Then you can fix the spelling, ignore it once or whitelist the offending word.
cspell also has a VSCode extension which picks up the configuration from a project and integrates with the "problems" panel, just like eslint, typescript and others.

> 3. yes that would be awesome, not sure how to deal with obfuscated / not-in-prod errors. Maybe those tests need to be in an if block?

Perhaps, I'll look into it.