@elektronik2k5 2020-04-25
> Unrelated, I was thinking making `observableRequiresReaction` the default, how'd you feel about that?

Sorry, I have no experience with any of the newer configuration options. Before MST I always worked in strict mode (is that `enforceActions: 'always'` nowadays?), but ever since switching to MST I haven't worked with bare mobx. :(

It is probably better to consult @Bnaya, @benjamingr and others.