@FredyC 2020-04-25
I think spellchecking in open source lib is a bit stretch. It could annoy contributors as any linter does and the benefit is very small imo. Who cares if there is a typo in a comment? Most people will understand it anyway.